requirements.txt
# ── Core ─────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0

# ── PDF Processing ──────────────────────────────────────────
PyMuPDF==1.24.0

# ── Image / Photo ───────────────────────────────────────────
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

# ── LLM Groq (optionnel — gratuit) ──────────────────────────
groq==0.13.0

# ── Upload handling ─────────────────────────────────────────
python-multipart==0.0.9
